requirements.txt
numpy<2
torch
transformers
openai-whisper
streamlit
pydub
soundfile
sentence-transformers
plotly
librosa
requests
zonos
